+++ src/main/java/com/github/dakusui/scriptiveunit/loaders/beans/ConstraintDefinitionImpl.java
@@ -1,8 +1,9 @@
-package com.github.dakusui.scriptiveunit.model.statement;
+package com.github.dakusui.scriptiveunit.loaders.beans;
 
 import com.github.dakusui.scriptiveunit.model.form.FormUtils;
 import com.github.dakusui.scriptiveunit.model.desc.ConstraintDefinition;
 import com.github.dakusui.scriptiveunit.model.session.Stage;
+import com.github.dakusui.scriptiveunit.model.statement.Statement;
 
 import java.util.List;
 
@@ -11,7 +12,7 @@ import static java.util.Objects.requireNonNull;
 public class ConstraintDefinitionImpl implements ConstraintDefinition {
   private final Statement statement;
 
-  public ConstraintDefinitionImpl(Statement statement) {
+  ConstraintDefinitionImpl(Statement statement) {
     this.statement = statement;
   }
